backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "celery-pool-asyncio>=0.2.0",
    "celery>=5.5.1",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pip>=25.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sendgrid>=6.11.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "uvicorn>=0.34.0",
]
